Fix symmetric-cluster example compilation issue some classes are moved to different packages
