+++ hornetq-core/src/main/java/org/hornetq/api/core/BroadcastEndpointFactoryConfiguration.java
@@ -19,7 +19,7 @@
 * Software Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA
 * 02110-1301 USA, or see the FSF site: http://www.fsf.org.
 */
-package org.hornetq.core.config;
+package org.hornetq.api.core;
 
 import java.io.Serializable;
