+++ hornetq-core/src/main/java/org/hornetq/api/core/UDPBroadcastGroupConfiguration.java
@@ -11,7 +11,7 @@
  * permissions and limitations under the License.
  */
 
-package org.hornetq.core.config;
+package org.hornetq.api.core;
 
 import org.hornetq.core.cluster.BroadcastEndpointFactory;
 import org.hornetq.core.cluster.UDPBroadcastEndpointFactory;
